+++ CHANGELOG.md
@@ -4,6 +4,8 @@ New features:
 - correct-material-command
 
 Changes:
+- XIT CONTS now displays a ! mark for actionable contracts
+- XIT CONTS now displays SHPT for contracts with shipment condition
 - Changed XIT calculator to https://desmos.com/scientific
 - Changed custom math evaluator to https://github.com/bugwheels94/math-expression-evaluator
 - Changed chart library to Chart.js, reducing the size of the plugin. Works on Firefox too!